@@ -1,9 +1,9 @@
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { Attachment } from '@wh-objects/attachment';
 import { HttpClient } from '@angular/common/http';
 
-const apiFiles = "https://wp.willsbach-handball.de/wp-json/wp/v2/media?_embed&search=downloads";
+const apiFiles = 'https://wp.willsbach-handball.de/wp-json/wp/v2/media?_embed&search=downloads';
 
 @Injectable()
 export class FileServiceService {
@@ -20,48 +20,48 @@ export class FileServiceService {
   }
 
   private getIconForType(attachment: Attachment): Attachment {
-    var icon: string;
+    let icon: string;
 
     switch (attachment.mime_type) {
-      case "application/pdf": {
-        attachment.iconColor = "red";
-        attachment.fileType = "PDF";
-        icon = "fa-file-pdf-o";
+      case 'application/pdf': {
+        attachment.iconColor = 'red';
+        attachment.fileType = 'PDF';
+        icon = 'fa-file-pdf-o';
         break;
       }
-      case "application/msword":
-      case "application/vnd.openxmlformats-officedocument.wordprocessingml.document":
+      case 'application/msword':
+      case 'application/vnd.openxmlformats-officedocument.wordprocessingml.document':
         {
-          attachment.iconColor = "blue";
-          attachment.fileType = "Word";
-          icon = "fa-file-word-o";
+          attachment.iconColor = 'blue';
+          attachment.fileType = 'Word';
+          icon = 'fa-file-word-o';
           break;
         }
-      case "application/msexcel":
-      case "application/vnd.ms-excel":
-      case "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet":
+      case 'application/msexcel':
+      case 'application/vnd.ms-excel':
+      case 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet':
         {
-          attachment.iconColor = "green";
-          attachment.fileType = "Excel";
-          icon = "fa-file-excel-o";
+          attachment.iconColor = 'green';
+          attachment.fileType = 'Excel';
+          icon = 'fa-file-excel-o';
           break;
         }
-      case "application/vnd.ms-powerpoint":
-      case "application/vnd.openxmlformats-officedocument.presentationml.presentation":
+      case 'application/vnd.ms-powerpoint':
+      case 'application/vnd.openxmlformats-officedocument.presentationml.presentation':
         {
-          attachment.iconColor = "orange";
-          attachment.fileType = "Power Point";
-          icon = "fa-file-powerpoint-o";
+          attachment.iconColor = 'orange';
+          attachment.fileType = 'Power Point';
+          icon = 'fa-file-powerpoint-o';
           break;
         }
       default: {
-        attachment.iconColor = "black";
-        attachment.fileType = attachment.mime_type.replace("application/", " ");
-        icon = "fa-file-o";
+        attachment.iconColor = 'black';
+        attachment.fileType = attachment.mime_type.replace('application/', ' ');
+        icon = 'fa-file-o';
       }
     }
 
-    attachment.icon = "fa " + icon;
+    attachment.icon = 'fa ' + icon;
     return attachment;
   }
 
