@@ -1,12 +1,13 @@
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { HttpClient } from '@angular/common/http';
-import { Ligue, Statistik, Club, StatGame, StatGoals } from '@wh-objects/hvw';
+import { Ligue, Club } from '@wh-objects/hvw';
 import { teamStatistics } from '@wh-objects/team-statistics';
 import { Globals } from '@wh-objects/globals';
 
 const baseUrl = 'https://spo.handball4all.de/service/if_g_json.php';
-const clubUrl = baseUrl + '?c=60&cmd=pcu&og=3&p=58';
+const tickerUrl = 'http://spo.handball4all.de/service/ticker.html?appid=&token=';
+const reportUrl = 'http://spo.handball4all.de/misc/sboPublicReports.php?sGID=';
 
 @Injectable()
 export class HvwService {
@@ -19,7 +20,7 @@ export class HvwService {
     this._liga = liga;
   }
 
-  private _allGames: string = '1';
+  private _allGames = '1';
   get allGames(): string {
     return this._allGames;
   }
@@ -27,40 +28,63 @@ export class HvwService {
     this._allGames = allGames;
   }
 
+  private _period: String = '58';
+  get period(): String {
+    return this._period;
+  }
+  set period(period: String) {
+    this._period = period;
+  }
+
   constructor(private http: HttpClient, private global: Globals) { }
 
   getLigueData(): Observable<Ligue> {
-    let url = this.buildUrlWithParam();
+    const url = this.buildUrlWithParam();
 
     return this.http.get<Ligue>(url).map(ligue => {
-      let data = ligue[0];
-      let scores = data.content.score;
+      const data = ligue[0];
+      const scores = data.content.score;
       scores.forEach(element => {
         element.difference = element.numGoalsShot - element.numGoalsGot;
       });
+      const games = data.content.futureGames.games;
 
-      let stats = new teamStatistics(this.global);
+      if (games !== undefined) {
+        games.forEach(element => {
+          if (element.live === true) {
+            element.tickerUrl = tickerUrl + element.gToken;
+          }
+          if (element.sGID !== undefined) {
+            element.pdfDL = reportUrl + element.sGID;
+          }
+        });
+      }
+
+
+      const stats = new teamStatistics(this.global);
       stats.calcStatistic(data);
       return data;
     });
   }
 
   getNextGames(): Observable<Ligue> {
-    let url = this.buildUrlWithParam();
-    return this.http.get<Ligue>(url).map(ligue => { return ligue[0]; });
+    const url = this.buildUrlWithParam();
+    return this.http.get<Ligue>(url).map(ligue => ligue[0]);
   }
 
   getClubData(): Observable<Club> {
+    const clubUrl = baseUrl + '?c=60&cmd=pcu&og=3&p=' + this._period;
+    console.log(clubUrl);
     return this.http.get<Club>(clubUrl).map(club => {
-      let data = club[0];
-      let classes = data.content.classes;
+      const data = club[0];
+      const classes = data.content.classes;
 
       classes.forEach(element => {
-        element.games.forEach(element => {
-          if (element.gGuestGoals === " ") { element.gGuestGoals = "0" }
-          if (element.gHomeGoals === " ") { element.gHomeGoals = "0" }
-          if (element.gGuestGoals_1 === " ") { element.gGuestGoals_1 = "0" }
-          if (element.gHomeGoals_1 === " ") { element.gHomeGoals_1 = "0" }
+        element.games.forEach(child => {
+          if (child.gGuestGoals === ' ') { child.gGuestGoals = '0'; }
+          if (child.gHomeGoals === ' ') { child.gHomeGoals = '0'; }
+          if (child.gGuestGoals_1 === ' ') { child.gGuestGoals_1 = '0'; }
+          if (child.gHomeGoals_1 === ' ') { child.gHomeGoals_1 = '0'; }
         });
       });
       return data;
