@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { Location, PlatformLocation } from '@angular/common';
 
-import { ReplaySubject } from 'rxjs/ReplaySubject';
-import 'rxjs/add/operator/do';
+import { ReplaySubject } from 'rxjs';
+
 
 
 @Injectable()
@@ -31,7 +31,7 @@ export class LocationService {
   }
 
   // TODO?: ignore if url-without-hash-or-search matches current location?
-  go(url: string|null|undefined) {
+  go(url: string | null | undefined) {
     if (!url) { return; }
     url = this.stripSlashes(url);
     if (/^http/.test(url) || this.swUpdateActivated) {
@@ -57,24 +57,24 @@ export class LocationService {
   }
 
   search() {
-    const search: { [index: string]: string|undefined; } = {};
+    const search: { [index: string]: string | undefined; } = {};
     const path = this.location.path();
     const q = path.indexOf('?');
     if (q > -1) {
       try {
-          const params = path.substr(q + 1).split('&');
-          params.forEach(p => {
-            const pair = p.split('=');
-            if (pair[0]) {
-              search[decodeURIComponent(pair[0])] = pair[1] && decodeURIComponent(pair[1]);
-            }
-          });
+        const params = path.substr(q + 1).split('&');
+        params.forEach(p => {
+          const pair = p.split('=');
+          if (pair[0]) {
+            search[decodeURIComponent(pair[0])] = pair[1] && decodeURIComponent(pair[1]);
+          }
+        });
       } catch (e) { /* don't care */ }
     }
     return search;
   }
 
-  setSearch(label: string, params: { [key: string]: string|undefined}) {
+  setSearch(label: string, params: { [key: string]: string | undefined }) {
     const search = Object.keys(params).reduce((acc, key) => {
       const value = params[key];
       return (value === undefined) ? acc :
@@ -130,8 +130,8 @@ export class LocationService {
     this.urlParser.href = relativeUrl;
 
     // don't navigate if external link or has extension
-    if ( anchor.href !== this.urlParser.href ||
-         !/\/[^/.]*$/.test(pathname) ) {
+    if (anchor.href !== this.urlParser.href ||
+      !/\/[^/.]*$/.test(pathname)) {
       return true;
     }
 
