@@ -1,21 +1,19 @@
 import { Component, OnInit } from '@angular/core';
-import { NgbCarouselConfig } from '@ng-bootstrap/ng-bootstrap';
 import { CarouselService } from '@wh-share/carousel.service';
 import { TeamWP } from '@wh-objects/wordPress';
 
 @Component({
   selector: 'app-img-carousel',
   templateUrl: './img-carousel.component.html',
-  styleUrls: ['./img-carousel.component.css']
+  styleUrls: ['./img-carousel.component.scss']
 })
 
 export class ImgCarouselComponent implements OnInit {
   teams: TeamWP[];
 
-  constructor(private transitionService: CarouselService, carouselConfig: NgbCarouselConfig) {
+  constructor(private transitionService: CarouselService) {
 
     this.transitionService.fetchTeams().subscribe(team => this.teams = team);
-    carouselConfig.interval = 5000;
   }
 
   ngOnInit() {
