@@ -2,6 +2,7 @@ import { Injectable } from '@angular/core';
 
 @Injectable()
 export class Globals {
-    clubName: string = 'TSV Willsbach';
+    clubName = 'TSV Willsbach';
+    readonly VAPID_PUBLIC_KEY = 'BIsi_9ZfOjP2Swwy45c9bTQR722-KW71YhH0_QfkFjEUj79HtAWObCsKy3W51jC30uvEOzUoZwKMPjbaKlBXaPk';
 
 }
