@@ -5,7 +5,16 @@ export class Sponsors {
     guid: Object;
     title: Object;
     content: Object;
-    media_details: Object;
+    media_details: {
+        width: number;
+        height: number;
+    };
     _embedded: Object;
+    source_url: string;
+    alt_text: string;
+
+    acf: {
+        sponsorUrl: string;
+    }
     thumbnail: string;
 }
