@@ -23,10 +23,10 @@ export class teamStatistics {
         statistik.homeGoalsShot.avarageGoals = statistik.homeGoalsShot.totalGoals / statistik.homeGoalsShot.gameAmount;
         statistik.homeGoalsGot.avarageGoals = statistik.homeGoalsGot.totalGoals / statistik.homeGoalsGot.gameAmount;
         statistik.homeWins.procentualWins = statistik.homeWins.wins / statistik.homeWins.gameAmount;
+
         statistik.awayGoalsShot.avarageGoals = statistik.awayGoalsShot.totalGoals / statistik.awayGoalsShot.gameAmount;
         statistik.awayGoalsGot.avarageGoals = statistik.awayGoalsGot.totalGoals / statistik.awayGoalsGot.gameAmount;
         statistik.awayWins.procentualWins = statistik.awayWins.wins / statistik.awayWins.gameAmount;
-
         // set value to 0 if undefinded (NaN)
         statistik.homeGoalsShot.avarageGoals = statistik.homeGoalsShot.avarageGoals || 0;
         statistik.homeGoalsGot.avarageGoals = statistik.homeGoalsGot.avarageGoals || 0;
