@@ -7,7 +7,7 @@ import { SeoService } from '@wh-share/seo.service';
 @Component({
   selector: 'app-post-detail',
   templateUrl: './post-detail.component.html',
-  styleUrls: ['./post-detail.component.css']
+  styleUrls: ['./post-detail.component.scss']
 })
 export class PostDetailComponent implements OnInit, OnDestroy {
 
