@@ -5,7 +5,7 @@ import { Sponsors } from '@wh-objects/sponsors';
 @Component({
   selector: 'app-sponsors',
   templateUrl: './sponsors.component.html',
-  styleUrls: ['./sponsors.component.css']
+  styleUrls: ['./sponsors.component.scss']
 })
 export class SponsorsComponent implements OnInit {
 
