@@ -1,3 +1,11 @@
 export const environment = {
-  production: true
+  production: true,
+  firebase: {
+    apiKey: "AIzaSyDN0fZC-xUP2VN7TgDTZCfKT6y-bzg0toI",
+    authDomain: "web-willsbach-handball.firebaseapp.com",
+    databaseURL: "https://web-willsbach-handball.firebaseio.com",
+    projectId: "web-willsbach-handball",
+    storageBucket: "web-willsbach-handball.appspot.com",
+    messagingSenderId: "546501708384"
+  }
 };
