@@ -4,5 +4,13 @@
 // The list of which env maps to which file can be found in `.angular-cli.json`.
 
 export const environment = {
-  production: false
+  production: false,
+  firebase: {
+    apiKey: 'AIzaSyDN0fZC-xUP2VN7TgDTZCfKT6y-bzg0toI',
+    authDomain: 'web-willsbach-handball.firebaseapp.com',
+    databaseURL: 'https://web-willsbach-handball.firebaseio.com',
+    projectId: 'web-willsbach-handball',
+    storageBucket: 'web-willsbach-handball.appspot.com',
+    messagingSenderId: '546501708384'
+  }
 };
